more test removal
